+++ configs/recognition/r2plus1d/r2plus1d_r34_8x8x1_180e_kinetics400_rgb.py
@@ -1,31 +1,7 @@
-# model settings
-model = dict(
-    type='Recognizer3D',
-    backbone=dict(
-        type='ResNet2Plus1d',
-        depth=34,
-        pretrained=None,
-        pretrained2d=False,
-        norm_eval=False,
-        conv_cfg=dict(type='Conv2plus1d'),
-        norm_cfg=dict(type='SyncBN', requires_grad=True, eps=1e-3),
-        conv1_kernel=(3, 7, 7),
-        conv1_stride_t=1,
-        pool1_stride_t=1,
-        inflate=(1, 1, 1, 1),
-        spatial_strides=(1, 2, 2, 2),
-        temporal_strides=(1, 2, 2, 2),
-        zero_init_residual=False),
-    cls_head=dict(
-        type='I3DHead',
-        num_classes=400,
-        in_channels=512,
-        spatial_type='avg',
-        dropout_ratio=0.5,
-        init_std=0.01))
-# model training and testing settings
-train_cfg = None
-test_cfg = dict(average_clips='prob')
+_base_ = [
+    '../../_base_/models/r2plus1d_r34.py', '../../_base_/default_runtime.py'
+]
+
 # dataset settings
 dataset_type = 'RawframeDataset'
 data_root = 'data/kinetics400/rawframes_train'
@@ -99,6 +75,9 @@ data = dict(
         data_prefix=data_root_val,
         pipeline=test_pipeline,
         test_mode=True))
+evaluation = dict(
+    interval=5, metrics=['top_k_accuracy', 'mean_class_accuracy'])
+
 # optimizer
 optimizer = dict(
     type='SGD', lr=0.1, momentum=0.9,
@@ -107,20 +86,8 @@ optimizer_config = dict(grad_clip=dict(max_norm=40, norm_type=2))
 # learning policy
 lr_config = dict(policy='CosineAnnealing', min_lr=0)
 total_epochs = 180
-checkpoint_config = dict(interval=5)
-evaluation = dict(
-    interval=5, metrics=['top_k_accuracy', 'mean_class_accuracy'])
-log_config = dict(
-    interval=20,
-    hooks=[
-        dict(type='TextLoggerHook'),
-        # dict(type='TensorboardLoggerHook'),
-    ])
+
 # runtime settings
-dist_params = dict(backend='nccl')
-log_level = 'INFO'
+checkpoint_config = dict(interval=5)
 work_dir = './work_dirs/r2plus1d_r34_8x8x1_180e_kinetics400_rgb/'
-load_from = None
-resume_from = None
-workflow = [('train', 1)]
 find_unused_parameters = False