+++ configs/recognition/slowonly/slowonly_r50_video_inference_4x16x1_256e_kinetics400_rgb.py
@@ -1,23 +1,9 @@
-model = dict(
-    type='Recognizer3D',
-    backbone=dict(
-        type='ResNet3dSlowOnly',
-        depth=50,
-        pretrained=None,
-        lateral=False,
-        conv1_kernel=(1, 7, 7),
-        conv1_stride_t=1,
-        pool1_stride_t=1,
-        inflate=(0, 0, 1, 1),
-        norm_eval=False),
-    cls_head=dict(
-        type='I3DHead',
-        in_channels=2048,
-        num_classes=400,
-        spatial_type='avg',
-        dropout_ratio=0.5))
+_base_ = ['../../_base_/models/slowonly_r50.py']
 
-test_cfg = dict(average_clips='prob')
+# model settings
+model = dict(backbone=dict(pretrained=None))
+
+# dataset settings
 dataset_type = 'VideoDataset'
 img_norm_cfg = dict(
     mean=[123.675, 116.28, 103.53], std=[58.395, 57.12, 57.375], to_bgr=False)