+++ configs/recognition_audio/audioonly/audioonly_r50_64x1x1_100e_kinetics400_audio_feature.py
@@ -1,21 +1,7 @@
-# model settings
-model = dict(
-    type='AudioRecognizer',
-    backbone=dict(
-        type='ResNetAudio',
-        depth=50,
-        pretrained=None,
-        in_channels=1,
-        norm_eval=False),
-    cls_head=dict(
-        type='AudioTSNHead',
-        num_classes=400,
-        in_channels=1024,
-        dropout_ratio=0.5,
-        init_std=0.01))
-# model training and testing settings
-train_cfg = None
-test_cfg = dict(average_clips='prob')
+_base_ = [
+    '../../_base_/models/audioonly_r50.py', '../../_base_/default_runtime.py'
+]
+
 # dataset settings
 dataset_type = 'AudioFeatureDataset'
 data_root = 'data/kinetics400/audio_feature_train'
@@ -75,6 +61,9 @@ data = dict(
         ann_file=ann_file_test,
         data_prefix=data_root_val,
         pipeline=test_pipeline))
+evaluation = dict(
+    interval=5, metrics=['top_k_accuracy', 'mean_class_accuracy'])
+
 # optimizer
 optimizer = dict(
     type='SGD', lr=2.0, momentum=0.9,
@@ -83,20 +72,9 @@ optimizer_config = dict(grad_clip=dict(max_norm=40, norm_type=2))
 # learning policy
 lr_config = dict(policy='CosineAnnealing', min_lr=0)
 total_epochs = 100
-checkpoint_config = dict(interval=5)
-evaluation = dict(
-    interval=5, metrics=['top_k_accuracy', 'mean_class_accuracy'])
-log_config = dict(
-    interval=1,
-    hooks=[
-        dict(type='TextLoggerHook'),
-        # dict(type='TensorboardLoggerHook'),
-    ])
+
 # runtime settings
-dist_params = dict(backend='nccl')
-log_level = 'INFO'
+checkpoint_config = dict(interval=5)
+log_config = dict(interval=1)
 work_dir = ('./work_dirs/' +
             'audioonly_r50_64x1x1_100e_kinetics400_audio_feature/')
-load_from = None
-resume_from = None
-workflow = [('train', 1)]