+++ docs/changelog.md
@@ -2,9 +2,22 @@
 
 ### master
 
+**Highlights**
+
+**New Features**
+
+- Support [imgaug](https://imgaug.readthedocs.io/en/latest/index.html) for augmentations in the data pipeline ([#492](https://github.com/open-mmlab/mmaction2/pull/492))
+
 **Improvements**
 
 - Support setting `max_testing_views` for extremely large models to save GPU memory used ([#511](https://github.com/open-mmlab/mmaction2/pull/511))
+- Add repo citation ([#545](https://github.com/open-mmlab/mmaction2/pull/545))
+
+**Bug and Typo Fixes**
+
+- Fix a bug about ava dataset validation ([#527](https://github.com/open-mmlab/mmaction2/pull/527).
+
+**ModelZoo**
 
 ### 0.10.0 (31/12/2020)